Break into two files to allow getting just a style object
Users can now `require('@eaze/css-reset/object')`
