+++ package.json
@@ -31,6 +31,7 @@
   "dependencies": {
     "cast-array": "~1.0.1",
     "insert-styles": "~1.0.0",
-    "to-css": "~1.2.1"
+    "to-css": "~1.2.1",
+    "xtend": "^4.0.1"
   }
 }
